data/requirements_w_preprocess.txt
bdpy==0.18
h5py==2.10.0
ipython==8.4.0
matplotlib==3.3.4
nibabel==3.2.2
nilearn==0.9.1
numpy==1.19.2
pandas==1.1.5
Pillow==9.2.0
progressbar33==2.4
pycocotools==2.0.4
pycocotools_windows==2.0.0.2
scikit_image==0.17.2
scikit_learn==1.1.1
scipy==1.5.2
skimage==0.0
tensorboardX==2.5.1
tensorflow==2.9.1
torch==1.10.2
torchvision==0.11.3
tqdm==4.64.0
